tests/e2e/pyproject.toml
[project]
name = "e2e"
version = "0.1.0"
description = "End-to-end tests for ai-stack"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-playwright>=0.5.0",
    "httpx>=0.27.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["."]
